Move view filters to unified namespace
