+++ wcfsetup/install/files/lib/system/listView/AbstractListView.class.php
@@ -13,8 +13,8 @@ use wcf\system\interaction\bulk\IBulkInteractionProvider;
 use wcf\system\interaction\IInteractionProvider;
 use wcf\system\interaction\InteractionContextMenuComponent;
 use wcf\system\interaction\InteractionContextMenuComponentConfiguration;
-use wcf\system\listView\filter\IListViewFilter;
-use wcf\system\listView\filter\exception\InvalidFilterValue;
+use wcf\system\view\filter\IViewFilter;
+use wcf\system\view\filter\exception\InvalidFilterValue;
 use wcf\system\request\LinkHandler;
 use wcf\system\WCF;
 use wcf\util\StringUtil;
@@ -65,7 +65,7 @@ abstract class AbstractListView
     private array $availableSortFields = [];
 
     /**
-     * @var array<string, IListViewFilter>
+     * @var array<string, IViewFilter>
      */
     private array $availableFilters = [];
 
@@ -470,13 +470,13 @@ abstract class AbstractListView
         return $this->availableSortFields;
     }
 
-    public function addAvailableFilter(IListViewFilter $filter): void
+    public function addAvailableFilter(IViewFilter $filter): void
     {
         $this->availableFilters[$filter->getId()] = $filter;
     }
 
     /**
-     * @param IListViewFilter[] $filters
+     * @param IViewFilter[] $filters
      */
     public function addAvailableFilters(array $filters): void
     {
@@ -491,7 +491,7 @@ abstract class AbstractListView
     }
 
     /**
-     * @return array<string, IListViewFilter>
+     * @return array<string, IViewFilter>
      */
     public function getAvailableFilters(): array
     {
