+++ wcfsetup/install/files/lib/system/listView/user/MembersListView.class.php
@@ -8,9 +8,9 @@ use wcf\data\user\UserProfileList;
 use wcf\event\listView\user\MembersListViewInitialized;
 use wcf\system\interaction\user\UserProfileInteractions;
 use wcf\system\listView\AbstractListView;
-use wcf\system\listView\filter\exception\InvalidFilterValue;
-use wcf\system\listView\filter\SelectFilter;
-use wcf\system\listView\filter\TextFilter;
+use wcf\system\view\filter\exception\InvalidFilterValue;
+use wcf\system\view\filter\SelectFilter;
+use wcf\system\view\filter\TextFilter;
 use wcf\system\listView\ListViewSortField;
 use wcf\system\user\storage\UserStorageHandler;
 use wcf\system\WCF;
