+++ wcfsetup/install/files/lib/system/view/filter/BooleanFilter.class.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace wcf\system\listView\filter;
+namespace wcf\system\view\filter;
 
 use wcf\data\DatabaseObjectList;
 use wcf\system\form\builder\field\AbstractFormField;
-use wcf\system\form\builder\field\TextFormField;
-use wcf\system\WCF;
+use wcf\system\form\builder\field\CheckboxFormField;
 
 /**
- * Filter for text columns.
+ * Filter for boolean columns.
  *
  * @author      Marcel Werk
  * @copyright   2001-2025 WoltLab GmbH
  * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since       6.2
  */
-class TextFilter extends AbstractFilter
+class BooleanFilter extends AbstractFilter
 {
     #[\Override]
     public function getFormField(): AbstractFormField
     {
-        return TextFormField::create($this->id)
-            ->label($this->languageItem);
+        return CheckboxFormField::create($this->id)
+            ->label($this->languageItem)
+            ->nullable();
     }
 
     #[\Override]
@@ -30,8 +30,14 @@ class TextFilter extends AbstractFilter
         $columnName = $this->getDatabaseColumnName($list);
 
         $list->getConditionBuilder()->add(
-            "{$columnName} LIKE ?",
-            ['%' . WCF::getDB()->escapeLikeValue($value) . '%']
+            "{$columnName} = ?",
+            [1]
         );
+    }
+
+    #[\Override]
+    public function renderValue(string $value): string
+    {
+        return '';
     }
 }
