+++ wcfsetup/install/files/lib/system/view/filter/LabelFilter.class.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace wcf\system\listView\filter;
+namespace wcf\system\view\filter;
 
 use wcf\data\DatabaseObjectList;
 use wcf\data\label\group\ViewableLabelGroup;
 use wcf\system\form\builder\field\AbstractFormField;
 use wcf\system\form\builder\field\label\LabelFormField;
-use wcf\system\listView\filter\exception\InvalidFilterValue;
+use wcf\system\view\filter\exception\InvalidFilterValue;
 
 /**
  * Filter that allows to filter a list view by labels.
