+++ wcfsetup/install/files/lib/system/view/filter/UserFilter.class.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace wcf\system\listView\filter;
+namespace wcf\system\view\filter;
 
 use wcf\data\DatabaseObjectList;
 use wcf\system\cache\runtime\UserRuntimeCache;
 use wcf\system\form\builder\field\AbstractFormField;
 use wcf\system\form\builder\field\user\UserFormField;
-use wcf\system\listView\filter\exception\InvalidFilterValue;
+use wcf\system\view\filter\exception\InvalidFilterValue;
 
 /**
  * Filter for columns that contain user ids.
